+++ PocketPulse/iOS/Application/Controllers/PocketPulseApp.swift
@@ -7,43 +7,27 @@
 import SwiftUI
 import SwiftData
 
-// Defines possible errors during app startup
-enum AppError: Error, LocalizedError {
-    case storage(message: String)
-    
-    var errorDescription: String? {
-        switch self {
-        case .storage(let message):
-            return "Storage Error: \(message)"
-        }
-    }
-    
-    var title: String {
-        return "App Error"
-    }
-}
-
 /// The main entry point for the PocketPulse application.
 @main
 struct PocketPulseApp: App {
     
     // MARK: - Dependencies
     
-    // The central DI container.
-    // We use @State to keep it alive for the app's lifecycle.
+    /// The central DI container.
+    /// We use @State to keep it alive for the app's lifecycle.
     @State private var appDI: AppDI
     
-    // Track any simplified error during startup to show an alert (like safe mode)
+    /// Track any simplified error during startup to show an alert (like safe mode)
     @State private var startupError: AppError?
     
-    // Defines if we are running in a memory-only fallback mode
+    /// Defines if we are running in a memory-only fallback mode
     @State private var isInMemoryFallback: Bool = false
     
     // MARK: - Initialization
     
     init() {
         // Safe container loading
-        let (container, error) = PocketPulseApp.makeSafeContainer()
+        let (container, error) = AppDI.makeSafeContainer()
         
         // Initialize AppDI with the safe container
         _appDI = State(initialValue: AppDI(container: container))
@@ -58,6 +42,8 @@ struct PocketPulseApp: App {
         NotificationManager.shared.requestAuthorization()
     }
     
+    // MARK: - Body
+    
     var body: some Scene {
         WindowGroup {
             AppContainerView(
@@ -70,38 +56,18 @@ struct PocketPulseApp: App {
             .environment(appDI.profileViewModel)
             // Error Alert for startup issues
             .alert(
-                "App Error", isPresented: Binding(
+                AppConstants.Strings.errorTitle, isPresented: Binding(
                     get: { startupError != nil },
                     set: { if !$0 { startupError = nil } }
                 )
             ) {
-                Button("OK", role: .cancel) { startupError = nil }
+                Button(AppConstants.Strings.ok, role: .cancel) { startupError = nil }
             } message: {
                 if isInMemoryFallback {
-                    Text("The app is running in Safe Mode due to a storage error. Changes may not be saved.\n\nError: \(startupError?.errorDescription ?? "Unknown")")
+                    Text(String(format: AppConstants.Strings.errorSafeModeMessage, startupError?.errorDescription ?? AppConstants.Strings.unknown))
                 } else {
-                    Text(startupError?.errorDescription ?? "Unknown Error")
+                    Text(startupError?.errorDescription ?? AppConstants.Strings.errorUnknown)
                 }
-            }
-        }
-    }
-    
-    // MARK: - Helper Methods
-    
-    /// Attempts to build the persistent container, falling back to in-memory on failure.
-    private static func makeSafeContainer() -> (ModelContainer, AppError?) {
-        do {
-            return (try AppDI.buildDefaultDBModelContainer(), nil)
-        } catch {
-            print("CRITICAL: Failed to load persistent store: \(error)")
-            let appErr = AppError.storage(message: error.localizedDescription)
-            
-            // Fallback to in-memory
-            if let memoryContainer = try? AppDI.buildInMemoryModelContainer() {
-                return (memoryContainer, appErr)
-            } else {
-                // If even in-memory fails, we can't run.
-                fatalError("CRITICAL: Could not create in-memory store fallback. Error: \(error)")
             }
         }
     }
