+++ PocketPulse/iOS/Core/Constants/AppConstants.swift
@@ -10,6 +10,18 @@ import SwiftUI
 
 struct AppConstants {
     
+    // MARK: - Strings
+    struct Strings {
+        static let errorTitle = "App Error"
+        static let errorStorageMessage = "Storage Error: %@"
+        static let errorSafeModeMessage = "The app is running in Safe Mode due to a storage error. Changes may not be saved.\n\nError: %@"
+        static let errorUnknown = "Unknown Error"
+        static let unknown = "Unknown"
+        static let ok = "OK"
+        static let criticalStorageFailure = "CRITICAL: Failed to load persistent store: %@"
+        static let criticalInMemoryFallbackFailure = "CRITICAL: Could not create in-memory store fallback. Error: %@"
+    }
+    
     // MARK: - Layout Dimensions
     struct Validation {
         static let passwordMinLength: Int = 8
@@ -80,9 +92,13 @@ struct AppConstants {
         
         static let paddingTagHorizontal: CGFloat = 6
         static let paddingTagVertical: CGFloat = 3
-        static let paddingTopNano: CGFloat = 2
-        static let paddingTiny: CGFloat = 2 // For "nano"
-        static let paddingTen: CGFloat = 10 
+        
+        
+        static let barWidth: CGFloat = 16
+        static let cornerRadiusNano: CGFloat = 4
+        static let blurRadiusSmall: CGFloat = 2
+        static let blurRadiusLarge: CGFloat = 20
+        static let shadowRadiusLarge: CGFloat = 10
     }
     
     // MARK: - Component Specific Sizes
@@ -135,11 +151,4 @@ struct AppConstants {
         static let standardDuration: Double = 0.3
         static let damping: Double = 0.7
     }
-}
-Extension AppConstants.Layout {
-    static let barWidth: CGFloat = 16
-    static let cornerRadiusNano: CGFloat = 4
-    static let blurRadiusSmall: CGFloat = 2
-    static let blurRadiusLarge: CGFloat = 20
-    static let shadowRadiusLarge: CGFloat = 10
 }
