+++ PocketPulse/iOS/Resources/Localizable.xcstrings
@@ -35,9 +35,6 @@
       "comment" : "Label for the y-axis in the bar chart displaying daily income and expense totals.",
       "isCommentAutoGenerated" : true
     },
-    "App Error" : {
-
-    },
     "bill_add_bill_button" : {
       "comment" : "Add Bill button label",
       "localizations" : {
@@ -1079,10 +1076,6 @@
         }
       }
     },
-    "OK" : {
-      "comment" : "The label of a button that dismisses an alert.",
-      "isCommentAutoGenerated" : true
-    },
     "PAID" : {
       "comment" : "A label indicating that a bill is paid.",
       "isCommentAutoGenerated" : true
@@ -1897,10 +1890,6 @@
           }
         }
       }
-    },
-    "The app is running in Safe Mode due to a storage error. Changes may not be saved.\n\nError: %@" : {
-      "comment" : "An alert message displayed when the app is running in Safe Mode, indicating that changes may not be saved.",
-      "isCommentAutoGenerated" : true
     },
     "transaction_add_expense_title" : {
       "comment" : "Title for add expense view",
